API-old/pyproject.toml
[tool.poetry]
name = "API"
version = "0.1.0"
description = ""
authors = ["loichu <loic@humbert-web.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
